Added component for sort
