+++ src/pages/main-page/main-page.tsx
@@ -1,6 +1,5 @@
 import { Helmet } from 'react-helmet-async';
 import { useState } from 'react';
-import classNames from 'classnames';
 import Container from '@components/container/container';
 import { Offer } from '@type/offer';
 import ListOffers from '@components/list-offers/list-offers';
@@ -10,6 +9,7 @@ import Location from '@components/location/location';
 import MainEmpty from '@components/main-empty/main-empty';
 import { useAppDispatch, useAppSelector } from '@hooks/index';
 import { selectLocation, setOffers } from '@store/action';
+import OfferSort from '@components/offer-sort/offer-sort';
 
 type TMainPageProps = {
   offers: Offer[];
@@ -20,11 +20,6 @@ function MainPage({ offers }: TMainPageProps): JSX.Element {
   const activeLocation = useAppSelector((state) => state.location);
   const selectedOffers = useAppSelector((state) => state.offers);
   const [selectedOffer, setSelectedOffer] = useState<Offer | null>(null);
-  const [isSortOpened, setIsSortOpened] = useState<boolean>(false);
-
-  const handleSortBlockClick = () => {
-    setIsSortOpened(!isSortOpened);
-  };
 
   const handleLocationChange = (location: string) => {
     dispatch(selectLocation(location));
@@ -59,21 +54,7 @@ function MainPage({ offers }: TMainPageProps): JSX.Element {
               <b className="places__found">
                 {selectedOffers.length} {selectedOffers.length === 1 ? 'place' : 'places'} to stay in {selectedOffers[0].city.name}
               </b>
-              <form className="places__sorting" action="#" method="get">
-                <span className="places__sorting-caption">Sort by</span>
-                <span className="places__sorting-type" tabIndex={0} onClick={handleSortBlockClick}>
-              Popular
-                  <svg className="places__sorting-arrow" width={7} height={4}>
-                    <use xlinkHref="#icon-arrow-select" />
-                  </svg>
-                </span>
-                <ul className={classNames('places__options places__options--custom', isSortOpened && 'places__options--opened')}>
-                  <li className="places__option places__option--active" tabIndex={0}>Popular</li>
-                  <li className="places__option" tabIndex={0}>Price: low to high</li>
-                  <li className="places__option" tabIndex={0}>Price: high to low</li>
-                  <li className="places__option" tabIndex={0}>Top rated first</li>
-                </ul>
-              </form>
+              <OfferSort />
               <ListOffers
                 offers={selectedOffers}
                 onOfferHover={setSelectedOffer}
